mod: Allow different title colors from Aqua
This will enable k9s themes to use a different title color other than
Aqua. The default still creates the traditional Aqua title.

The Styles.BackgroundColor could be overridden too, but it never
gets used since it will be overridden by the background color of the
form which is exposed.

This is related to derailed/k9s#1149